# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SmartFarm AI API
  description: Complete API specification for SmartFarm AI - Poultry Farm ERP & Billing System
  version: 1.0.0
  contact:
    name: SmartFarm AI Team
servers:
  - url: http://localhost:8080/api
    description: Local Development Server
  - url: https://your-cloud-run-url/api
    description: Production Server (GCP Cloud Run)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SettlementRequest:
      type: object
      required:
        - customerId
        - amount
        - paymentMethod
        - transactionType
      properties:
        customerId:
          type: string
          format: uuid
        amount:
          type: number
          format: decimal
        paymentMethod:
          type: string
          enum: [CASH, CHECK, TRANSFER, ESEWA, CARD]
        notes:
          type: string
        transactionDate:
          type: string
          format: date-time
        transactionType:
          type: string
          enum: [RECEIPT, PAYOUT]
...
paths:
...
  /finance/settlements:
    post:
      tags:
        - Finance
      summary: Create payment settlement
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettlementRequest'
      responses:
        '200':
          description: Settlement processed
          content:
            application/json:
              schema:
                type: object
  /finance/customers/{customerId}/unpaid-sales:
    get:
      tags:
        - Finance
      summary: Get unpaid sales for customer
      security:
        - bearerAuth: []
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of unpaid sales
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      format: uuid
                    totalAmount:
                      type: number
                    paidAmount:
                      type: number
                    balance:
                      type: number
                    date:
                      type: string
                      format: date-time
  /finance/report:
    get:
      tags:
        - Finance
      summary: Get profit/loss report
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Financial report
          content:
            application/json:
              schema:
                type: object
  /finance/transactions:
    get:
      tags:
        - Finance
      summary: Get transaction report
      security:
        - bearerAuth: []
      parameters:
        - name: customerId
          in: query
          schema:
            type: string
            format: uuid
        - name: fromDate
          in: query
          schema:
            type: string
            format: date-time
        - name: toDate
          in: query
          schema:
            type: string
            format: date-time
        - name: paymentStatus
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Transaction report
          content:
            application/json:
              schema:
                type: object
  /finance/ledger:
    get:
      tags:
        - Finance
      summary: Get unified ledger
      security:
        - bearerAuth: []
      parameters:
        - name: customerId
          in: query
          schema:
            type: string
            format: uuid
        - name: fromDate
          in: query
          schema:
            type: string
            format: date-time
        - name: toDate
          in: query
          schema:
            type: string
            format: date-time
        - name: paymentStatus
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Unified ledger
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object

  # Dashboard
...